backend/tryon/requirements.txt
requests>=2.31.0
boto3>=1.34.0
Pillow>=10.0.0